requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Login==0.6.2
pytube==15.0.0